refactor(server): move routes into a subdirectory
